+++ src/utils/show.jl
@@ -166,21 +166,15 @@ Base.show(io::IO, x::RRTMGPI.AbstractRRTMGPMode) =
 Base.show(io::IO, x::Setups.RCEMIPIIProfile) =
     parseable_show_with_fields_no_type_header(io, x)
 
-import ClimaCore: Fields, Spaces
 function Base.show(
     io::IO, ::MIME"text/plain", x::Setups.ColumnInterpolatableField,
 )
-    # Extract z grid from the wrapped column field
-    z = Fields.coordinate_field(x.f).z
-    nz = Spaces.nlevels(z)
-    zmin, zmax = extrema(z)
-    val_eltype = eltype(x.f)
-    # These are fixed by the constructor
-    interp_str = "Linear"
-    extrap_str = "Flat"
+    z = x.itp.xsource
+    nz = length(z)
+    val_eltype = eltype(x.itp.fsource)
     print(io,
-        "ColumnInterpolatableField(Nz=$nz, z∈[$zmin, $zmax], value_eltype=$val_eltype, ",
-        "interpolation=$interp_str, extrapolation=$extrap_str)",
+        "ColumnInterpolatableField(Nz=$nz, z∈[$(first(z)), $(last(z))], value_eltype=$val_eltype, ",
+        "interpolation=Linear, extrapolation=Flat)",
     )
 end